feat: Add ToolBar component
Added a new Tool the. ToolBar is for displaying a navigation bar at top of the page. It includes icons for the home, user, new, and hashtag pages. The selected icon is highlighted with a background color. The ToolBar component also handles navigation when the user clicks on an icon.