feat: Add maxMatch function and word_list
This commit adds the maxMatch function and word_list module to the project. The maxMatch function implements a word segmentation algorithm using a maximum matching approach. The word_list module contains a set of words that can be used for segmentation. This new feature will enhance the user experience and improve the accuracy of word segmentation.