feat: Add SwitchButton component
This commit adds a new component called SwitchButton, which is used to toggle between two options. The component takes a callback function as a prop, which is called when the button is clicked. The callback function receives an index parameter that indicates the selected option (0 for "Top" and 1 for "Latest"). The selected option is stored in the component's state and can be accessed using the dataType prop. The SwitchButton component is used in the ThreadsList component