feat: Add ButtonTools component and refactor SwitchButton component
This commit adds a new ButtonTools component and refactors the existing SwitchButton component to use the ButtonTools component. The ButtonTools component allows the user to switch and triggers a callback function when the switch is toggled. The SwitchButton component has been deleted as it is no longer needed. This commit also includes some minor changes to the SigList and ThreadsList components.