feat: Add support for displaying announcement threads on mobile
This commit adds support for displaying announcement threads on mobile devices. The Thread component now accepts an optional prop "isAnnouncement" which, when set to true, will render the thread as an announcement with a different styling. The changes include adding a new CSS class "threadAnnouncement" to the Thread.module.scss file and updating the Thread.tsx file to conditionally render the announcement thread based on the value of the "isAnnouncement" prop